PyCharm_Projects/ResizeNSaveImageToBTM/requirements.txt
future==0.18.2
numpy==1.19.2
Pillow==7.2.0
py2exe==0.9.2.2
six==1.15.0
